Make sure rgb data is float as well

+++ scripts/data_setup.py
@@ -71,7 +71,7 @@ class LettuceSegDataset(Dataset):
 
     def __getitem__(self, index):
         # Retrieve image and mask, should be np.array for albumentations.transforms
-        img = np.array(Image.open(self.img_paths[index]))
+        img = np.array(Image.open(self.img_paths[index]), dtype=np.float32)
         if self.mask_paths[index].endswith(".json"):
             mask = utils.binary_poly2px(self.mask_paths[index]).astype(np.float32)
         else:
